+++ app/javascript/packs/hbw.js
@@ -1,4 +1,3 @@
-import 'babel-polyfill';
 import 'url-polyfill';
 import 'whatwg-fetch';
 import 'modulejs';